+++ CHANGELOG.md
@@ -373,3 +373,4 @@
 * Define `Stack` interface with `push` and `pop` methods
 * Define new type synonymous: `:=:=>`
 * Define experimental `Prepend` verb for `Morphable` ability
+* Remove `Insertable` ability for datastructures